+++ kmk/boards/converter/keebio/nyquist_r2.py
@@ -1,7 +1,7 @@
 import board
 
-from kmk.consts import DiodeOrientation
-from kmk.mcus.circuitpython_usbhid import KeyboardConfig as _KeyboardConfig
+from kmk.matrix import DiodeOrientation
+from kmk.keyboard_config import KeyboardConfig as _KeyboardConfig
 
 
 class KeyboardConfig(_KeyboardConfig):
